third_party/WebKit/LayoutTests/media/picture-in-picture/disable-picture-in-picture.html

<!DOCTYPE html>
<title>Test disable Picture-in-Picture</title>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="resources/picture-in-picture-helpers.js"></script>
<body></body>
<script>
promise_test(t => {
  const video = document.createElement('video');
  video.disablePictureInPicture = true;
  return promise_rejects(t, 'InvalidStateError',
      requestPictureInPictureWithTrustedClick(video));
}, 'Request Picture-in-Picture rejects if disablePictureInPicture is true');

promise_test(t => {
  const video = document.createElement('video');
  return requestPictureInPictureWithTrustedClick(video)
  .then(() => {
    video.disablePictureInPicture = true;
    assert_equals(document.pictureInPictureElement, null);
  });
}, 'pictureInPictureElement is unset if disablePictureInPicture becomes true');

promise_test(t => {
  const video = document.createElement('video');
  return requestPictureInPictureWithTrustedClick(video)
  .then(() => {
    video.disablePictureInPicture = false;
    assert_equals(document.pictureInPictureElement, video);
  });
}, 'pictureInPictureElement is unchanged if disablePictureInPicture becomes false');


promise_test(t => {
  const pipVideo = document.createElement('video');
  const notPipVideo = document.createElement('video');
  return requestPictureInPictureWithTrustedClick(pipVideo)
  .then(() => {
    notPipVideo.disablePictureInPicture = true;
    assert_equals(document.pictureInPictureElement, pipVideo);
  });
}, 'pictureInPictureElement is unchanged if another disablePictureInPicture ' +
   'becomes true for another video');
</script>
